requirements.txt
arrow==1.2.1
asgiref==3.4.1
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cookiecutter==1.7.3
Django==3.2.9
django-cleanup==5.2.0
django-tailwind==2.2.2
django-widget-tweaks==1.4.9
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
MarkupSafe==2.0.1
Pillow==8.4.0
poyo==0.5.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
